Change in the initial state
At the Nth position of the array for the initial state the correct is to set if n<max_N: ... 0 else ...1 according to the notation in the report. However, for some reason that I don't understand, the results are the same in both ways. The code in quotes is an example for you to run